// app.client/src/api/models/TagIntermediateVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TagIntermediateVM = {
    tagIntermediateId?: string | null;
    tagId?: string | null;
    tagName?: string | null;
    name?: string | null;
    category?: string | null;
    categoryId?: string | null;
};